==> NWPtools/personal_folders/dnk/03_Filling Table 1A_dnk_input.Rmd <==
---
title: "Annex 3 - R script to construct Table 1A by country"
author: Joel Vigneau, with the collaboration of Matt Elliott, Jon Elson, Kirsten Birch   Hakansson,
  Marie Storr-Paulsen, Nuno Prista, Katja Ringhdal, Lies Vansteenbrugge,   Sieto Verver
date: "October 2016"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Filling of the EU-MAP table 1A requires to report on country shares of landings and shares of EU TAC when relevant, for all the stocks listed in table 1A of the EU-MAP Regulation (EU Decision 1254/2016). This process necessitates to gather information on landings and EU TAC from an official database, namely EUROSTAT for EU landings and MARE/FIDES for EU TAC. 

Two datasets were added to complete the references, (1) the Nephrops FU landings provided by ICES and (2) the Mediterranean and Black Sea landings fisgures put together by 2016 RCM Mediterranean and Black Sea.

First of all, the datasets listed above contain information from all EU Member States, which means that the script has the potential to be used by all Member States, and by STECF for control of the NWP submitted for 2017. 


## Setting the parameters for the analysis
The variables needed for the work are the working directory, the country code (2-letter code) and the reference years 

```{r echo=FALSE}

input_path_common <- "Q:/scientific-projects/eu-data-collection/Work_Plan/2020/scripts/gits/RCGs/NWPtools/common/input"
input_path_dnk <- "Q:/scientific-projects/eu-data-collection/Work_Plan/2020/scripts/gits/RCGs/NWPtools/personal_folders/dnk/input"

output_path <- "Q:/scientific-projects/eu-data-collection/Work_Plan/2020/scripts/gits/RCGs/NWPtools/personal_folders/dnk/"

```

```{r warning=FALSE}
library(tidyr)
library(stringr)
library(reshape)
library(dplyr)

CTRY <- 'DK'
refYears <- 2016:2018
```
## Importing the datasets
The list of datasets are the following :

1. Landings and TAC shares files: 
  + EUROSTAT landings files: http://ec.europa.eu/eurostat/web/fisheries/data/database
  + MARE/FIDES TAC file: https://webgate.ec.europa.eu/fides/index.cfm
  + ICES Nephrops fishery units landings per country for 2015
  + RCM Mediterranean and Black Sea 2016 landings compilation
  
  
```{r}
DF27  <- read.table(file.path(input_path_common, "fish_ca_atl27.tsv"),header=TRUE, sep='\t', as.is=TRUE)  # Atlantic NE
DF21  <- read.table(file.path(input_path_common, "fish_ca_atl21.tsv"),header=TRUE, sep='\t', as.is=TRUE)  # Atlantic NW
DF34  <- read.table(file.path(input_path_common, "fish_ca_atl34.tsv"),header=TRUE, sep='\t', as.is=TRUE)  # Atlantic CE
DF41  <- read.table(file.path(input_path_common, "fish_ca_atl41.tsv"),header=TRUE, sep='\t', as.is=TRUE)  # Atlantic SW
DF47  <- read.table(file.path(input_path_common, "fish_ca_atl47.tsv"),header=TRUE, sep='\t', as.is=TRUE)  # Atlantic SE
DF51  <- read.table(file.path(input_path_common, "fish_ca_ind51.tsv"),header=TRUE, sep='\t', as.is=TRUE)  # Indian W
DF37  <- read.table(file.path(input_path_common, "fish_ca_atl37.tsv"),header=TRUE, sep='\t', as.is=TRUE)  # Med&BS

DF <- rbind.data.frame(DF27, DF21)
DF <- rbind.data.frame(DF, DF34)
DF <- rbind.data.frame(DF, DF37)
DF <- rbind.data.frame(DF, DF41)
DF <- rbind.data.frame(DF, DF47)
DF <- rbind.data.frame(DF, DF51)

TAC <- read.csv(file.path(input_path_common,'EU opening quota 2018.csv'), header=TRUE, sep=';',as.is=TRUE) 
NEP <- read.csv(file.path(input_path_common,'Nephrops landings 2015.csv'), header=TRUE, sep=';', as.is=TRUE)
patch_land_codIIIa <- read.table(file.path(input_path_common, 'patch_codIIIa_20191009.txt'), sep = '\t', header = TRUE, as.is = TRUE)
MED <- read.csv(file.path(input_path_common,'RCM MED landings.csv'), header=TRUE, sep=';', as.is=TRUE)

```
2. Reference tables:
  + EuroStat Geo.def: full names of countries
  + ASFIS file : FAO species naming and coding
  + Linkage table mirroring EU-MAP Table 1A naming of species and stock area, and lining to EUROSTAT and MARE/FIDES species and area naming

```{r}
GEO     <- read.table(file.path(input_path_common,'geo.def'),header=TRUE,sep=";", as.is=TRUE)
ASFIS   <- read.table(file.path(input_path_common,'ASFIS_sp_Feb_2016.txt'), header=TRUE, sep="\t", as.is=TRUE)
table1A <- read.csv(file.path(input_path_dnk,'EUMAP_Table1A_Linkage_EUROSTAT and EC_TAC_dnk.csv'), sep=';',header=TRUE, as.is=TRUE)
```

## data.frame preparation
The country names are matching between GEO and TAC data.frame, except for UK, so the following lines enables the full match.

```{r warning = FALSE}
TAC$StockID <- paste(TAC$Species.Code,TAC$Area.Code,sep="")
TAC$Level.Description[substring(TAC$Level.Description,1,3) %in% 'U.K'] <- 'United Kingdom'
```
The TAC dataset is well structured and thus ready for the analysis
```{r warning = FALSE}
head(TAC,3)
names(GEO)[2] <- "Country"
GEO$geo <- toupper(GEO$geo) #2-letter code should be in capitals
SRG <- strsplit(as.character(DF$species.fishreg.unit.geo.time),split=",")
SRG.m <- matrix(unlist(SRG), ncol=4, byrow=TRUE)
coln <- sapply(refYears, function(x) which(grepl(x,names(DF))))
DFT <- data.frame(X3A_CODE = toupper(SRG.m[,1]), area = toupper(SRG.m[,2]), geo = SRG.m[,4],
Y1 = DF[,coln[1]], Y2 = DF[,coln[2]], Y3 = DF[,coln[3]])
DFM <- merge(DFT, GEO, all.x=TRUE)
DFM$Y1 <- as.numeric(str_remove_all(as.character(DFM$Y1), "[bcdefinpzsu]")) #Removing characthers in amount
DFM$Y2 <- as.numeric(str_remove_all(as.character(DFM$Y2), "[bcdefinpzsu]")) #Removing characthers in amount
DFM$Y3 <- as.numeric(str_remove_all(as.character(DFM$Y3), "[bcdefinpzsu]")) #Removing characthers in amount
DFM <- DFM[!is.na(DFM$Country),]
DFM <- merge(DFM, ASFIS[,c(3:6)], all.x=TRUE)
```
Let's have a look at the workable structure of EuroStat dataset. Note that Y1, Y2 and Y3 are the 3-year period demanded, and the presence of NA's. The assumption made here (further in the Construction of the table section) is to exclude NA from the average, i.e. like if MS had omitted to report, instead of a NA which would mean 0. The confusion comes because lots of 0 are reported in EuroStat (implicitely meaning that NA is not a 0). This point may be subject of a STECF agreement or suggestion for modification.
```{r warning = FALSE}
head(DFM,3)
NEP <- merge(NEP, GEO, all.x=TRUE)
NEP$geo[is.na(NEP$geo)] <- 'UK'
NEP2 <- data.frame(X3A_CODE='NEP', geo=NEP$geo, area=NEP$Stock, Y1=round(NEP$TotalLanding.in.kg/1000,0), Y2=NA, Y3=NA, Country=NEP$Country, Scientific_name='Nephrops norvegicus', English_name='Norway lobster', French_name='Langoustine')
```
A look at the Nephrops dataset on the same format as EuroStat dataset, so they can be merged
```{r warning = FALSE}
head(NEP2)
DFM <- rbind.data.frame(DFM, NEP2)
MEDA <- merge(MED, ASFIS[,c(3,4,5,6)], by.x='Species', by.y='Scientific_name', all.x=TRUE)
MEDA <- tidyr::gather(MEDA,"Country","n",4:13) 
MEDAG <- merge(MEDA, GEO, all.x=TRUE)
```
and a look at the Mediterranean dataset
```{r warning = FALSE}
head(MEDAG,3)
MED <- data.frame(X3A_CODE=MEDAG$X3A_CODE, geo=MEDAG$geo, area=MEDAG$Area, Y1=round(MEDAG$n,0),
                   Y2=NA, Y3=NA, Country=MEDAG$Country, Scientific_name=MEDAG$Species, English_name=MEDAG$English_name,
                   French_name=NA)
DFM <- rbind.data.frame(DFM, MED)
```

## Construction of the table
```{r}
T1A <- data.frame()
for (i in 1:nrow(table1A)) {
  
	ctry2 <- GEO$Country[GEO$geo %in% CTRY]
	reg <- strsplit(as.character(table1A$areaBis[i]), split=',')
	
	if (table1A$region[i] %in% 'Mediterranean and Black Sea') {
	    DT <- DFM[DFM$Scientific_name %in% table1A$latinName[i] & tolower(DFM$area) %in% tolower(paste(reg[[1]], collapse=',')),]
	  } else {
	    
	    ting<-unlist(strsplit(table1A$latinName[i], split=","))
	    DT <- DFM[DFM$Scientific_name %in% ting & tolower(DFM$area) %in% tolower(reg[[1]]),]
	    
	    	  }
	
	if (table1A$latinName[i]=="Gadus morhua" & (table1A$area[i] %in% c("IIIaN","IIIaS")) & sum(refYears == 2016:2018)==3)
    {
     a<-melt(DT[c("geo","Y1","Y2","Y3")])
     a$id<-paste(a$geo,a$variable)
     patch_land_codIIIa$id <- paste0(patch_land_codIIIa$geo, " Y", patch_land_codIIIa$year_seq)
     a<-data.frame(a, b=patch_land_codIIIa[,table1A$area[i]][match(a$id, patch_land_codIIIa$id)]*a$value)
     DT[,4:6]<-matrix(data=a$b, ncol=3)
    }
	
	DT$MOY <- apply(DT[,4:6],1,mean,na.rm=TRUE)
	#RFMO <- 'ICES'
	if (substring(table1A$region[i],1,3) %in% 'Med') RFMO <- 'GFCM' 
	
	T1 <- data.frame(MS=CTRY, refYears=paste(min(refYears),'-',max(refYears),sep=""),spp=table1A$latinName[i],region=table1A$region[i],
		RFMO=table1A$RFMO[i], area = table1A$area[i],select=NA, landings=NA, TAC=NA,shareLanding=NA,Thresh='N', Comments=NA, Comments2=NA)
	ind <- which(DT$geo %in% CTRY)
	if (length(ind)>0) {
		T1$landings <- sum(DT$MOY[DT$geo %in% CTRY],na.rm=TRUE)
		T1$shareLanding <- T1$landings/sum(DT$MOY[DT$geo != 'EU28'], na.rm=TRUE)
		#T1$shareLanding <- T1$landings/sum(DT$MOY, na.rm=TRUE)
		} else {
		T1$landings <- 0
		T1$shareLanding <- 0
		}
	
	## TAC
	ind.ct<-NULL
	if (!(table1A$FIDES_stockID[i] %in% 'No TAC')) {
	  	aa<-strsplit(as.character(table1A$FIDES_stockID[i]),split=',')[[1]]
		TACi <- TAC[TAC$StockID %in% aa,]
		if (length(aa)>1) 
			TACi <- aggregate(list(Initial.Quantity = TACi$Initial.Quantity), 
				by=list(Level.Code=TACi$Level.Code, Level.Description=TACi$Level.Description), sum)
		ind.ct <- TACi$Initial.Quantity[which(TACi$Level.Description %in% ctry2)]
		ind.eu <- TACi$Initial.Quantity[which(TACi$Level.Code %in% 'EEC')]
		if (length(ind.ct) == 1) T1$TAC <- ind.ct/ind.eu
		T1$Comments<-NA
		TT <- tapply(TACi$Initial.Quantity, TACi$Level.Description,sum,na.rm=TRUE)/TACi$Initial.Quantity[TACi$Level.Code %in% 'EEC']
		TT <-  TT[names(TT) %in% GEO$Country]  #Keep only the EU countries to calculate the 25% rule
		if (!(is.na(T1$TAC)) & T1$TAC <0.1 & T1$TAC>0) 	T1$Comments <- sum(TT[which(TT<0.1)])
		if (!(is.na(T1$Comments)) & T1$Comments >=.25)	{
			print(T1)
			print(TT[TT<.1])
			cat('\n')
		}
	}
	#Add-on Sept 2019 Joel
	if (length(ind.ct)>0) {
	  if (!is.na(ind.ct))  {
	  T1$Comments2 <- paste('FIDES Initial.Quantity =',ind.ct)
	}}
	##
	T1A <- rbind.data.frame(T1A, T1)
	T1A$Thresh <- as.character(T1A$Thresh)
	#Threshold ruling
	# T1A$Thresh[T1A$TAC >=.1 & T1A$landings >=200] <- 'M'  #rule (a) & (c)
	# T1A$Thresh[is.na(T1A$TAC) & T1A$shareLanding >=.1 & T1A$landings >=200] <- 'M' #rule (b) & (c)
	# T1A$Thresh[T1A$TAC <.1 & T1A$Comments >=.25] <- 'C'   # 25% rule, sampling to be coordinated between MS
	#Threshold ruling specified like the EU Reg
	T1A$Thresh[T1A$TAC <.1] <- 'Y'  #rule (a)
	T1A$Thresh[is.na(T1A$TAC) & T1A$shareLanding <.1] <- 'Y' #rule (b)
	T1A$Thresh[T1A$landings < 200] <- 'Y'  #rule (c)
	T1A$Thresh[T1A$TAC <.1 & T1A$Comments >=.25] <- 'N'   # 25% rule, sampling to be coordinated between MS
}
```



##Formatting
```{r}
T1B <- T1A
T1B$landings <- round(T1B$landings,0)
T1B$landings[T1B$landings == 0] <- '-'
T1B$TAC <- paste(round(100*T1B$TAC,0),'%',sep='')
T1B$TAC[T1B$TAC %in% c('NA%','NaN%','Inf%')] <- '-'
T1B$shareLanding <- paste(round(100*T1B$shareLanding,0),'%',sep='')
T1B$shareLanding[T1B$shareLanding %in% c('NA%','NaN%','Inf%')] <- '-'
T1B$Thresh[T1B$landings %in% '-' & T1B$TAC %in% '-'] <- T1B$shareLanding[T1B$landings %in% '-' & T1B$TAC %in% '-'] <- '-'
ind <- which(T1B$Comments>.0) #Changed, so I know what the figures means
T1B$Comments <- paste(round(100*T1B$Comments,0),'%',sep='')
T1B$Comments[T1B$Comments<.25] <- '-'
T1B$Comments[T1B$Comments %in% c('NA%','NaN%','Inf%')] <- '-'
T1B$Comments[ind] <- paste('Sum of MS shares <10% = ',T1B$Comments[ind],sep='')
T1B$Comments[!ind]<-'-'
T1B$select <- '-'
T1B$select[T1B$Thresh %in% c('N')] <- 'Y'
T1B$select[T1B$Thresh %in% 'Y'] <- 'N'
T1B[T1B$RFMO %in% c('ICCAT','IOTC','WCPFC') & T1B$landings > 0,c('select','Thresh')] <- c('Y','N')
#T1B[T1B$spp %in% 'Anguilla anguilla' & T1B$landings > 0, c('select','Thresh')] <- c('Y','N')

T1B$select <- ifelse(T1B$spp %in% 'Anguilla anguilla' & T1B$landings > 0, "Y", T1B$select)
T1B$Thresh <- ifelse(T1B$spp %in% 'Anguilla anguilla' & T1B$landings > 0, "N", T1B$Thresh)

T1B[T1B$spp %in% 'Nephrops norvegicus' & !(grepl('TAC', T1B$area)) & T1B$landings < 0,'TAC'] <- '-' #Changed a bit, so it don't overwrites TAC in 3a
levels(T1B$refYears) <- c(levels(T1B$refYears), '2015')
T1B[T1B$spp %in% 'Nephrops norvegicus' & !(grepl('TAC', T1B$area)) & T1B$landings < 0,'refYears'] <- '2015' #Not totally correct ...
T1B[T1B$RFMO %in% 'GFCM', 'refYears'] <- '2015'
```

```{r dnkInfo}
#Added so we can use the output to create table 1B & C
names(T1B)
names(table1A)
T1B$Comments_comb <- ifelse(T1B$Comments != "-" & !(is.na(T1B$Comments2)), paste(T1B$Comments, T1B$Comments2, sep = " | "), 
                            ifelse(T1B$Comments == "-" & !(is.na(T1B$Comments2)), T1B$Comments2,
                                   ifelse(T1B$Comments != "-" & is.na(T1B$Comments2), T1B$Comments, "")))


T1C <- merge(select(distinct(T1B), -Comments, -Comments2), select(distinct(table1A), -Comments), by.x = c("region","spp","area","RFMO"), by.y = c(c("region","latinName","area","RFMO")))
names(T1C)

T1C <- mutate(T1C, speciesIncluded = spp, spp = reportingName, Comments = Comments_comb)
names(T1C)

T1C<-T1C[,c("MS","refYears","spp","region","RFMO","area","select","landings","TAC","shareLanding","Thresh","Comments","areaBis","speciesIncluded")]

```


## Export of Table 1A
the rule sum of quotas for coutries <10% (less or more than 25%) is noted in the comments column
```{r}
ind <- order(as.character(T1B$region), as.character(T1B$RFMO), as.character(T1B$spp), as.character(T1B$area))
write.table(T1C[ind,], file=paste(output_path, CTRY,'_table1A_filled_dnk.csv',sep=''), sep=';',row.names=FALSE, quote=FALSE)
```
